refactor: controller 제거
