+++ src/models/index.js
@@ -1,2 +1,1 @@
-export * from "./racingGame.js";
 export * from "./car.js";
